When unhooking, verify that the function wasn't changed
If for some reason the function was changed, e.g. if somebody else
placed a hook on top of ours, don't restore the original bytes over it,
which will remove the other hook and will likely cause a crash when that
other library will remove its hook. Instead, leak the allocated
trampoline and patch it such that it will jump straight to the original
code.