[SLIMDETOURS] Small fixes and improvements (#6)
* detour_is_imported: Check only current bitness magic in PE header

The code proceeds to access IMAGE_OPTIONAL_HEADER of the current
bitness, so it makes little sense to proceed if different bitness was
matched, which is also quite unlikely.

* detour_is_imported: Add check for NumberOfRvaAndSizes

The IMAGE_DIRECTORY_ENTRY_IAT data directory entry might be missing.

* SlimDetoursTransactionCommit: Remove unnecessary pointer check

The pointer is being accessed just before the NULL check.

* SlimDetoursDetach: Remove unnecessary detour_skip_jmp call

The pointer to the trampoline was set by us on attach.