Chain multiple hooks to the same function in a single transaction
Currently, it's assumed that there can only be a single operation per
target function in a single transaction. Attaching multiple hooks to the
same function doesn't work properly, and removing one hook and setting
another doesn't work either.

This change addresses this limitation as following:

* Hook addition operations are done before hook removal operations
  during the commit. Hook addition relies on the disassembly of the
  original function as it was seen before the commit started. Hook
  removal operations may modify the function and invalidate this
  assumption.
* Consecutive hook addition operations for the same target are now
  chained.

Note that hook removal operations handle the case of the function
changing during the commit. In this case, the hook will be put in bypass
mode, resulting in a correct removal of the hook but in a memory leak
for the allocated trampoline. This case can be optimized in the future.